+++ src/util.js
@@ -16,6 +16,7 @@ import {
   OPEN_REQUEST_STATUSES,
   DCB_HOLDINGS_RECORD_ID,
   DCB_INSTANCE_ID,
+  DCB_USER_LASTNAME,
 } from './constants';
 
 export function getFullName(user) {
@@ -96,3 +97,5 @@ export function renderOrderedPatronBlocks(patronBlocks) {
 }
 
 export const isDCBItem = (item) => item.instanceId === DCB_INSTANCE_ID && item.holdingsRecordId === DCB_HOLDINGS_RECORD_ID;
+
+export const isDCBUser = (user) => user.lastName === DCB_USER_LASTNAME;